[IMP] accounting_restrict_date: simplify logic
