+++ accounting_restrict_date/models/__init__.py
@@ -1,1 +1,3 @@
-from . import order_date
+from . import purchase
+from . import sale
+from . import invoice
